Allow adding multiple files for monitoring (windows)
Had to rename System::RemoveDirectory() to RemovePath(), because
apparently on windows, RemoveDirectory is a built-in #define. Ugh.
Renamed MakeDirectory() to MakePath() to be consistent.

Not ready to be merged to master yet, some of this isn't really RAII
compliant yet, and I think I'm taking the mutex in the file monitor
implementations too much.