A tale of SFINAE and std::hash
So, some compilers (*cough* vc++ *cough*) implement std::hash in a way
that makes it not work with SFINAE tests. C++14 extends std::hash with
this specialization:

http://en.cppreference.com/w/cpp/utility/hash
In addition to the above, the standard library provides specializations
for all (scoped and unscoped) enumeration types. These may be (but are
not required to be) implemented as
std::hash<std::underlying_type<Enum>::type>)

In this specialization, some compilers add a struct-level static_assert
which makes sure the given type is indeed an enum. This static_assert
prevents the SFINAE check from working in std::enable_if. This SO
question explains it more:

https://stackoverflow.com/questions/12753997/check-if-type-is-hashable

As a workaround, the standard specifies that std::hash<T *> should be
defined by the compiler, so we can fall back to that in
String::getValue() if operator<< is not defined. Callers may still
specialize std::hash with their own definition, and getValue() will use
it. Otherwise, it will call the generic version defined by the compiler
instead.