Remove the DownCast() function from fly.h I intended this function to prevent needing run-time checks by using compile-time checks instead; it performs a static (compile-time) check when downcasting a shared_ptr of type A to type B to ensure A is a parent of B. Then it does a static cast instead of a dynamic cast (which has run-time checking), assuming that the cast will be safe.
But, that static check doesn't prevent this scenario:

    class Parent
    {
    public:
        virtual ~Parent() = default;
    };

    class ChildA : public Parent
    {
    };

    class ChildB : public Parent
    {
    };

    int main()
    {
        std::shared_ptr<ChildA> spChildA = std::make_shared<ChildA>();

        std::shared_ptr<Parent> spParent = spChildA;

        std::shared_ptr<ChildB> spChildB = DownCast<ChildB>(spParent);
        std::cout << static_cast<bool>(spChildB) << std::endl;

        return 0;
    }

Since spParent is really holding an object of type ChildA, static casting the
object to type ChildB is heap corruption. But DownCast() allows this because
we're telling it to convert from Parent to ChildB.

So, remove the DownCast() helper function. Instead, use std::static_pointer_cast
when we KNOW the cast is valid. Otherwise, use std::dynamic_pointer_cast and
check if the returned shared_ptr is valid before using it.