Mocked Linux system calls for unit tests The largest hole in libfly unit testing is testing failed system calls. GCC allows mocking system calls with the --wrap linker option.
Add calls to the build system to make it easy for unit tests to declare system
calls that should be mocked. Add test/mock library for implementing the mocked
system calls and making run-time decisions for whether the mocked or real
system call will be invoked.

As proof of concept, implement mock system calls for testing the PathMonitor
class.

Will need to research an equivalent feature for Windows.