Add connect() and getsockopt() mock system calls and fix found bugs The getsockopt() call is used for asynchronous connects, to verify the socket is healthy after an async connect completes. This is called in the socket manager thread, not the main thread. So, MockSystem needs to be thread-safe; the unit test occasionally failed dued to the main thread's MockSystem object going out of scope while the socket manager thread was inside a call to getsockopt() (then, really, inside a call to MockEnabled()).
The socket managers were incorrectly erasing closed sockets from the list of
asynchronous sockets in the handleSocketIO() loop. Normally it's okay to do
something like "it = vec.erase(it)", so long as you keep track of whether or
not you need to increment the iterator later on. That wasn't being tracked
correctly. Rather than try to get that right, just delete the closed sockets
later on, outside of the loop.