Fix errors on windows
JSON parser must read characters as ints. They can't be read as char on Windows because std::isspace apparently throws an exception if the value is not in the range [-1, 255]; a signed char will be negative when reading a unicode character. They also can't be read as unsigned char because std::ifstream::get(c) does not compile if 'c' is unsigned.

Fix crash dereferencing an iterator when it points at the end of the stream (didn't need it really).

In types_test, can just read large unicode as UTF-8 still (actually get errors about UTF-16 not defined).