Clean up interface for creating sockets
Don't need separate methods for TCP & UDP, just pass in the Protocol enum.

Move Protocol enum out of Socket class so that it can be forward declared.

Change all "if tcp / else if udp" branches to a switch statement with no default, so errors will be created for unhandled cases.