Clean up socket interface (#6)
Some methods took string addresses/hostnames as parameters, others took
numeric addresses. For flexibility, provide both for each of those methods.

A bunch of methods don't need to be repeatedly defined for each socket
implementation. Move them from the SocketImpl class to the Socket class.

The methods that take in a bool to store whether the operation would
block or is complete do not need to be public.