Make socket test faster by not using huge messages (#10)
The point of using a really large message was to force packet fragmentation.
This can be mocked instead. Add a mock system flag for send() to enable mocking
EWOULDBLOCK. If set, send all but the last byte of the string, setting errno to
EWOULDBLOCK. On subsequent calls, send the last byte through.

Also fix some bad return value checks in the Window's socket implementation
(should be checking for SOCKET_ERROR instead of -1).