New task interface (#12)
The old interface of having classes implement the Runner interface was kind of
clunky. It results in lots of threads often doing nothing but sleeping, and it
also made unit tests take forever since it was hard to make the runners wake up
sooner than however long they wanted to sleep for.

The new task manager owns pool of threads that any class can post tasks to once
there is actually work to do. Task runners may be created from the task manager
to govern execution order of posted tasks. For unit testing, there are also
"waitable" task runners to allow the main thread to block until a specific task
is complete (instead of sleep_for(5s) and hoping that's the right amount of time
to sleep).

This isn't really complete in that some classes still need to be converted to be
fully task-friendly. E.g. the socket manager class - instead of having methods
like WaitForCompletedSend(), the completion of those sends should be hooked into
the task manager to invoke a registered completion callback.