Use C++17 std::variant in JSON class (#27)
Using std::variant is nicer as a type-safe union. Don't need to maintain type
(can query type at run time), and not limited to POD and pointer types.

This also increases stack size in the unit tests for Windows, otherwise a SEH
exception is raised in unit tests with very deep JSON nesting. This is probably
because object and array types are used directly in the std::variant, whereas
the union had pointers and dynamically allocated objects and arrays on the heap.