Generalize fly::String for std::string and std::wstring (#32)
Lo-and-behold, std::filesystem::path uses std::wstring as a string type
on Windows. Everywhere that std::filesystem::path is used now passes in
std::string. This seems to causes Windows to sporadically fail Release
unit tests. I suspect this is undefined behavior, though I'm surprised
it doesn't result in a compile error.

Step 1 to being able to use std::wstring on Windows is to generalize
fly::String. If not for the streaming methods in fly::String, it could
actually accept std::u16string and std::u32string as well. So in the
future, maybe specializations without streaming can be defined for those
types.