Replace FLY_DECLARATION_TESTS with DeclarationTraits (#37)
Having this feature exist as a macro was a bit unnerving. Replace it with an
extensible templated struct. Extend this struct in BasicStringTraits<> to
define a trait for whether operator<< is declared for a type. Use that trait
in BasicString<>::stream() - if the type is not streamable, stream its address
instead.