Optimizations round 5: Memory (chunks) As an intermediate step, I was reading the entire file at once into a buffer before encoding the stream. Of course, this is bad for really large files. Instead, read the file in (currently 1MB) chunks. Encode each chunk as if it is an entire stream. This seems to have little effect on the compression ratio.
Since streams are encoded in chunks, it follows that streams must also be
decoded in chunks. The decoder decodes symbols until either the stream has
reached EOF or the chunk size has been reached. Interestingly, this seems to
have had a small improvement on decoding speed.

A TODO is that the chunk size should be encoded into the output stream. This
way, if the chunk sizes changes between builds, the decoder can still read old
files.