Optimizations round 6: Prefix tables (decoder)
Converting decoded Huffman codes to a Huffman tree and navigating through that
tree bit-by-bit to decode symbols was quite slow. The code was simple, but each
bit in the encoded stream resulted in many branches and pointer hopping. Change
the decoder to form prefix tables instead of a Huffman tree.

This brings decoding time with -02 on a 100MB file down from ~3.6 seconds to
~1.4s.