Add tests for coding files, and fix issues found First, the base Coder class was opening output files incorrectly.
Second, the HuffmanCoder had bugs related to data types. There are 256 ASCII
codes. If we're *really* unlucky, we'll have a file to encode in which all
ASCII codes appear equally and very frequently (which is true for test.bin).
The result is terrible compression, as all Huffman codes will have length 8.
Then, when encodeCodes() encodes the code length counts, it tried to store 256
in a single byte - which can be at most 255. It unfortunately has to be changed
to encode code lengths in 2 bytes. Some other fallout occurs.