Simpler JSON parser (#112)
The existing JSON parser was just a mess. It maintained a state machine and
several other boolean states in an attempt to iteratively parse JSON files.
It was pretty fragile though and hard to read & follow.

This new parser is recursive, which isn't ideal, but it at least doesn't
blow out the stack in any of the existing test cases. The main improvement
is that it doesn't suck to look at.