Make Json array conversion operator more generic
The array-like STL containers do not all define methods like insert()
the same. So the existing method of converting a Json instance to an
array-like container only worked for some STL containers. This adds a
trait that wraps around each array-like container's insert / push_back
methods to allow the Json class to convert to all of them.