Support Json iterator access (#118)
Initial support for iterating over Json instances. Only objects and arrays are
supported. It would be nice to support strings, but the question becomes how
to define things like operator*. For objects and arrays, this method can return
a Json reference. But for strings, it would have to return a char.

A main goal of this iterator is to prevent undefined behavior. This includes
incrementing past the end iterator, and detecting iterators to objects that
no longer exist.