Add key/value accessors for Json object iterators
Iterators for Json objects only store an object's value. It'd be nice to store
a std::pair<key, value>, but that isn't feasible yet. Providing a separate key
accessor will suffice for now.

So for a Json object, iterator.key() is analogous to iterator->first, and
iterator.value() is analogous to iterator->second.