Move unicode processing to String
Rather than performing unicode processing in the Json class, it is better
to do this work in String. This allows for handling unicode processing in
all string types (std::string, std::wstring, std::u16string, std::u32string).

As for being a full-fledged unicode implementation, the new String method
is pretty limiting and incomplete. But it serves as a good starting point
to make unicode processing generic.