Move unicode processing to String (#129)
Rather than performing Unicode processing in the Json class, it is better
to do this work in String. This allows for handling Unicode processing in
all string types (std::string, std::wstring, std::u16string, std::u32string).

As for being a full-fledged Unicode implementation, the new String method
is pretty limiting and incomplete. But it serves as a good starting point
to make Unicode processing generic.