Custom reporter for Catch2 unit tests
Catch2 doesn't report test case or section names. For tests which have
a fair amount of output, this makes it pretty difficult to differentiate
tests. Define a custom reporter to pretty-print these names.

It is a bit more verbose than I'd like (sections that only serve as
common setup steps also get reported). But it's fine for now.