Reduce code generation in JSON operator==
The definition of operator== contains a sequence of constexpr if
statements, and returns false if none of them are hit. Put that
return in an else block so that it isn't counted as missing code
coverage when one of the if constexpr blocks is entered. In that
case, it will be unreachable.