CaptureStream should flush the stream before and after capture
This ensures CaptureStream captures exactly what is streamed when it is
in scope. If the stream is not flushed before capturing, it may contain
unflushed data that was streamed earlier. If the stream is not flushed
after capturing, the file the captured stream is redirected to may not
contain all of the desired output.

Also make opening the file the captured stream is redirected to required
to succeed, rather than silently ignoring errors.