Move numeric literal suffixes to the fly namespace
To avoid polluting the global namespace, put the numeric literal suffixes
in the fly namespace. But to discourage callers from importing all of the
fly namespace (using namespace fly), also nest the suffixes in inline
namespaces.

The standard going forward will be to inline all literal suffixes in the
fly;:literals namespace, and further in a second inline namespace specific
to only that file (fly::literals::numeric_literals in this case). This is
the practice used by the STL as well.