Ensure JAR contains compiled anonymous inner classes
If a Java class contains an anonymous inner class, then javac will
create multiple class files for that class. This entirely breaks the
build system's current strategy of maintaining a 1-to-1 mapping of Java
files to class files.

I'm not yet sure of a foolproof way to determine the exact class files
that javac will create. Until then, compile a target to a subdirectory
of $(CLASS_DIR), and include that entire subdirectory in the JAR.