Overhaul of logging system
Previously, it was only valid to create a single logger, and that logger
could only log to a file. Or, for console logging, not creating a logger
at all would log to the console.

Any number of loggers may now be created, though there is still a notion
of a default logger for the LOG* macros. There are also now log sinks to
receive log points and stream them however they wish. There are built-in
sinks for console and file logging.

Lastly, there is now a logger registry whose job is (will be) to manage
all created loggers. This will allow creating named loggers, and then
retrieving those loggers by name at will. For now, it only manages the
default logger.