Improved task runner interface
Previously, to post a task to a task runner, the caller had to have a
subclass of fly::Task overriding its run() method. This was pretty ugly
because in most instances, that task would then have friend access to
whatever class owned it to do the actual work.

Task runners now accept any callable type as tasks, e.g. lambdas or
std::function. This allows tasks to be created much more easily without
having to give out friend access.

Posting a task also now requires a source location object to indicate
where the task was posted from. This is a bit cumbersome, but this will
allow for tracking tasks, as well as lets the testing waitable task
runners work without RTTI.