Logger no longer needs to store log points in a queue
The fields required to create a log can be forwarded directly to the
task. Ideally, a log point could be created in Logger::log and moved to
the posted task. But something about the wrapped task semantics in the
task runner cause a compile error due to the log points being move-only.