Task runner API for receiving task results
Add methods to receive the result of a task in a reply task. Currently
the reply task is posted on the same task runner on which the original
task was posted. I think it'd be nicer if the reply was posted on the
task runner (if any) on which the caller that posted the task was
currently on.