Fix flakiness of SequencedTaskRunner::maybe_post_task
The following order of events occurred semi-regularly on Azure (or some
similar order):

1. Thread A (user thread) posts a task an enters maybe_post_task().
2. Thread A is halted before checking the atomic task-running flag.
3. Thread B (task manager thread) enters task_complete() and then
   maybe_post_task().
4. Thread B sees no task is running. It takes the task posted by
   thread A.
5. Task posted by Thread A completes.
6. Thread A is allowed to continue running. No task is running.
7. Thread A tries to take the task it queued, but is no longer there.
8. The call to post_task() returns false even though the task ran.

This isn't so bad in that the posted task was skipped, but the return
value is errantly false. The fix is to not put the task onto the pending
queue until after checking if it can be run immediately.