Replace escaped JSON character(s) in-place
Rather than creating a copy of the input string in Json::validate_string,
replace the escaped control and Unicode characters in-place.

In a release build, this reduces the time to run the all_unicode.json
test from ~1.6 seconds to ~1.0 seconds. In a debug build, time reduced
from ~17.6 seconds to ~16.6 seconds.