Do not accumulate parsed JSON strings in a stream
Rather than using std::basic_stringstream to accumulate the parsed JSON
string, just append to a string object directly. The performance gain
here is likely due to avoiding a copy when calling stream.str().

In a release build, this reduces the time to run the all_unicode.json
test from ~1.0 seconds to ~0.4 seconds. In a debug build, time reduced
from ~15.1 seconds to ~12.7 seconds.