Improve Base64 decoder performance by error checking only when needed
The implementation of decoding a single chunk can be specialized to only
handle padding symbols on the very last chunk of the encoded stream. For
all other chunks, the decoder can effieciently check if any symbol is
invalid or padding by examining just the MSB.

With this change, Base64 decoding performance is improved by about 60ms
on the enwik8 file (about a 20% improvement).