Fix x86 memory corruption in Windows path monitor
There was some sort of memory corruption with the IOCP used in the
Windows PathMonitorImpl on x86 builds. Not 100% sure, but I think it was
due to the ULONG_PTR cast here:

    ::CreateIoCompletionPort(m_handle, iocp, (ULONG_PTR)m_handle, 0);

The m_handle variable is type HANDLE, which is void*, and will be 32-bit
in x86 builds. Casting to ULONG_PTR treats the value as 64-bit, so it
seems likely that some 64-bit value is written or read from a 32-bit
allocation.

The fix is that the IOCP isn't even needed. The polling thread can
rely on GetOverlappedResult instead of GetQueuedCompletionStatus (which
requires an IOCP). This change has the added benefit that more than one
monitored directory can be handled per poll loop.