Add string utility for lexical analysis of string literals
This is a thin wrapper around std::basic_string_view to assist with
lexical analysis of strings. Provides some basic peeking and consuming
methods, and ensures the caller cannot move past the end of the string.

This is currently in the fly::detail namespace because it's only use so
far is a rewrite of the string formatter. It could be promoted up to the
fly namespace if needed later.