String formatting overhaul: Format according to std::format specs
The previous commit added support for (nearly) fully compliant parsing
of the new std::format style of string. This implements the actual
formatting of strings according to this format.

The following features are not yet supported:
1. Nested replacement fields (see previous commit).
2. Center alignment - will be treated as type-default alignment for now.
3. Wide-to-narrow character validation - a simple static_cast is used.

To support all formatting options, a few helper facets are added. These
are a facet to use a space in place of a positive sign, and a facet to
format a value as a binary value (because there isn't a standard binary
manipulator such as std::bin).

This also drops BasicString::create_hex_string. This method was needed
because the old formatting implementation did not support fixed-width
hexadecimal formatting, but now e.g. format("{:2x}", 0x4) will suffice.