Add fly::net::TcpSocket and fly::net::UdpSocket
These will serve as a much better replacement for fly::Socket. The old
socket class doesn't support IPv6 at all and is rather weakly typed.
That is, for example, it's entirely possible to attempt TCP operations
on a UDP socket. This would fail at runtime, but now with the new socket
classes, it's not even possible to attempt.

The old socket design was such that there was a virtual base socket
interface, and concrete implementations for each platform. In this new
design, platform-dependent system calls are abstracted into a "socket
operations" file, which share a common interface that the socket classes
may then interact with.

Currently, while these sockets may be configured to be non-blocking,
there isn't a service for monitoring non-blocking sockets yet. There's
also still some room for improvement regarding configuring sockets. For
example, to create a listening socket, users will need to create a
TcpSocket and invoke bind/listen/accept on that socket. It would be nice
to have a concrete "ListenSocket" class to hide these details.