Implement a new asynchronous socket service
Adds fly::net::SocketService for polling asynchronous sockets for IO
readiness. This is a replacement for fly::SocketMananger. The new system
continues to use ::select() for polling, just to get the infrastructure
in place for now. Ideally, this will be updated to use ::epoll() on
POSIX systems and IO completion ports on Windows.