Convert formatting of all types to fly::Formatter specializations
Including user-defined types, rather than formatting parameters with an
operator<< overload, rely on specializations of fly::Formatter. At least
for now, this also removes default formatting of strongly types enums,
because it's kind of a pain to set up a specialization for those types.

To prevent trying to format types for which there isn't a fly::Formatter
specialization, a C++20 concept is introduced to ensure the structure is
defined for a given type.