Update JSON for Modern C++ version 3.10.2
Note that JSON for Modern C++ now uses the following attribute on the
parse() function: __attribute__((__warn_unused_result__))

While Clang allows using an explicit cast to (void) to ignore the result
of this invocation, GCC does not. The result must be stored, and then
the result can be ignored.