Validate replacement fields for user-defined types at compile-time
While the result cannot be stored at compile-time, replacement fields of
user-defined types can be parsed for validation.

A couple of things were moved around to accomplish this. Parsing up to
and including the colon occurs before parsing standard or user-defined
replacement fields. Parsing the closing brace of standard fields now
occurs in FormatSpecifier to align with |parse| requirements.