Re-use compile-time parsed replacement fields for user-defined types
For user-defined types which:

    1. Are derived from a standard formatter.
    2. Do not define a custom parsing method.

The formatter does not need to reparse the replacement field at runtime.
Instead, the data that was parsed at compile time can be copied into the
derived formatter.

This has the added benefit of allowing nested replacement fields to work
with these kinds of user-defined types. Previously, reparsing the field
would cause BasicFormatParseContext::next_position() to roll past the
end of the format parameters.