Do not use assertion __VA_ARGS__ directly in constant expressions
By passing __VA_ARGS__ to an consteval method, FLY_ASSERT had implicitly
required the captured variables be able to be constant evaluated. This
fails when trying to capture e.g. a class member variable with errors
such as:

    note: implicit use of 'this' pointer is only allowed within the
          evaluation of a call to a 'constexpr' member function

Instead, create a helper to create the array of captured value names
directly. This will always be able to be constant evaluated.